backend/requirement.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
openpyxl==3.1.2
pydantic==2.8.2